# docs/swagger.yaml
basePath: /
definitions:
  models.CreateEmployee:
    properties:
      hired_date:
        example: "2024-06-01"
        type: string
      name:
        example: John Doe
        type: string
      position:
        example: Software Engineer
        type: string
      salary:
        example: 60000
        type: integer
    type: object
  models.Employee:
    properties:
      created_at:
        example: "2024-06-10T12:00:00Z"
        type: string
      hired_date:
        example: "2024-06-01"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John
        type: string
      position:
        example: Software Engineer
        type: string
      salary:
        example: 60000
        type: integer
      updated_at:
        example: "2024-06-10T12:00:00Z"
        type: string
    type: object
  utility.ErrorResponse:
    properties:
      error:
        example: err
        type: string
      message:
        example: An error occurred
        type: string
      status:
        example: 500
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is an API for managing employees.
  title: Employee Management API
  version: "1.0"
paths:
  /employees:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Employee'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
      summary: List employees
      tags:
      - employees
    post:
      consumes:
      - application/json
      parameters:
      - description: Employee Info
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.CreateEmployee'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
      summary: Create a new employee
      tags:
      - employees
  /employees/{id}:
    delete:
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
      summary: Delete an employee
      tags:
      - employees
    get:
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
      summary: Retrieve an employee by ID
      tags:
      - employees
    put:
      consumes:
      - application/json
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Employee Info
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.CreateEmployee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ErrorResponse'
      summary: Update an employee
      tags:
      - employees
swagger: "2.0"